server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
deprecation==2.1.0
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-contrib-python-headless==4.9.0.80
packaging==25.0
pillow==11.3.0
polars==1.33.0
postgrest==1.1.1
psutil==7.0.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
realtime==2.7.0
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.14.0
tifffile==2025.8.28
# CPU-only PyTorch + Torchvision
torch==2.2.0+cpu
torchvision==0.17.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
ultralytics==8.3.196
ultralytics-thop==2.0.17
urllib3==2.5.0
websockets==15.0.1
